project: update gitignore o ignore database file
