+++ .env
@@ -1,3 +1,6 @@
 SECRET_KEY=YOUR_SECRET_KEY
+
 VAPID_DIR=YOUR_VAPID_DIRECTORY_PATH
 VAPID_CLAIM_EMAIL=YOUR_VAPID_CLAIM_EMAIL
+
+SQLITE_PATH=PATH_TO_YOUR_DB_FILE
