--- test.py
@@ -1,1 +1,0 @@
-from F
